+++ library/QATools/QATools/HtmlElements/Proxy/TypifiedElementCollectionProxy.php
@@ -14,6 +14,8 @@ namespace QATools\QATools\HtmlElements\Proxy;
 use QATools\QATools\HtmlElements\Element\AbstractTypifiedElementCollection;
 use QATools\QATools\HtmlElements\Element\ITypifiedElement;
 use QATools\QATools\PageObject\Element\AbstractElementCollection;
+use QATools\QATools\PageObject\ElementLocator\IElementLocator;
+use QATools\QATools\PageObject\IPageFactory;
 
 /**
  * Class for lazy-proxy creation to ensure, that TypifiedElementCollection are
@@ -25,6 +27,22 @@ use QATools\QATools\PageObject\Element\AbstractElementCollection;
  */
 class TypifiedElementCollectionProxy extends TypifiedElementProxy
 {
+
+	/**
+	 * Initializes proxy for AbstractTypifiedElementCollection.
+	 *
+	 * @param IElementLocator $locator      Element selector.
+	 * @param IPageFactory    $page_factory Page factory.
+	 * @param string          $name         Name of the element.
+	 */
+	public function __construct(IElementLocator $locator, IPageFactory $page_factory, $name)
+	{
+		if ( !$this->className ) {
+			$this->className = '\\QATools\\QATools\\HtmlElements\\Element\\AbstractTypifiedElementCollection';
+		}
+
+		parent::__construct($locator, $page_factory, $name);
+	}
 
 	/**
 	 * Offset to set.